# calculate q