--- a/src/Amica.jl
+++ b/src/Amica.jl
@@ -5,6 +5,9 @@ module Amica
 	using Distributions
 	using SpecialFunctions
     using ProgressMeter
+    using LoopVectorization
+    using AppleAccelerate
+    
     include("types.jl")
     include("helper.jl")
     include("likelihood.jl")