--- a/src/helper.jl
+++ b/src/helper.jl
@@ -34,10 +34,11 @@ end
 #taken from amica_a.m
 #L = det(A) * mult p(s|θ)
 function logpfun(x,rho)
-	return  (.-abs.(x).^rho .- log(2) .- loggamma.(1+1/rho))
+	return @inbounds -AppleAccelerate.pow(abs.(x), repeat([rho], length(x))) .- log(2) .- loggamma.(1 + 1 / rho)
 end
 
+
 #taken from amica_a.m
-function ffun(x,rho)
-	return rho .* sign.(x) .* abs.(x) .^(rho.-1)
+function ffun(x::AbstractArray{T, 1}, rho::T) where {T<:Real}
+	return @inbounds rho .* sign.(x) .* AppleAccelerate.pow(abs.(x), repeat([rho - 1], length(x)))
 end