--- a/src/helper.jl
+++ b/src/helper.jl
@@ -33,14 +33,14 @@ end
 
 #taken from amica_a.m
 #L = det(A) * mult p(s|θ)
-function logpfun(x,rho)
-	return @inbounds -optimized_pow(abs.(x), rho) .- log(2) .- loggamma(1 + 1 / rho)
+function logpfun(rho, y_rho)
+	return @inbounds - y_rho .- log(2) .- loggamma(1 + 1 / rho)
 end
 
 
 #taken from amica_a.m
-function ffun(x::AbstractArray{T, 1}, rho::T) where {T<:Real}
-	return @inbounds rho * sign.(x) .* optimized_pow(abs.(x), rho - 1)
+@views function ffun(x::AbstractArray{T, 1}, rho::T) where {T<:Real}
+	return @inbounds copysign.(optimized_pow(abs.(x), rho - 1), x) .* rho
 end
 
 # optimized power function for different cpu architectures