- Update typings to be more type-stable
- Reduce memory allocation and implement other performance improvements
- add array dimensions
- re-enable multi threading
- make Calculate_Q type stable
- use AppleAccelerate in certain places
- add intelvectormath
- minor
- more optimized methods
- minor
- log
- overflow fixed, debug statements added, comparson to fortran/matlab semi-automated
- fixes, make fp local
- update project.toml
- precalculate y^rho
- things are faster now
- optimize more sums
- fix sphering LL
- speed improvement from 272µs to 640ns
- insane additional 100x speed improvement in calculate_y
- this should work, but havent tested -oops
- added missing broadcast to first minus, 340 to 230 µs